+++ cypress/e2e/System/AboutPage/Slogan.cy.ts
