+++ cypress/e2e/System/faqPage.cy.ts
